Finish Robot Name PHP.
I think this is kind of a mess. I'm still trying to wrap my head around classes in PHP.
There has to be a better way to generate the name, but this works for now.
I'd like to revisit this one once I have a better grasp of this language